knownProgrammer/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cloudinary==1.39.1
Django==4.2.11
django-bootstrap-v5==1.0.11
django-multiselectfield==0.1.12
django-phonenumber-field==7.3.0
python-dotenv==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.2.1
